categorys for manager

+++ system/modules/Ui/Controllers/DataManagerController.php
@@ -72,12 +72,53 @@ class DataManagerController extends Controller {
         ob_end_clean();
         $result->send();
     }
+    function loadCategorysAction() {
+        $result = new Server\Result();
+        ob_start();
+        if (strpos($_GET['modelName'], ':')) {
+            $raw = explode(':', $_GET['modelName']);
+            $modelName = $raw[0];
+            $id = $raw[1];
+            $model = $modelName::get($id);
+        } else {
+            $modelName = $_GET['modelName'];
+            $id = null;
+            $model = null;
+        }
+        if (!empty($_GET['params'])) {
+            $params = $_GET['params'];
+            if (!empty($params['relation'])) {
+                $relations = $modelName::relations();
+                $modelName = $relations[$params['relation']]['model'];
+            }
+        } else {
+            $params = [];
+        }
+        $dataManager = new Ui\DataManager($modelName, $_GET['managerName']);
+        $dataManager->drawCategorys();
+        $result->content = ob_get_contents();
+        ob_end_clean();
+        $result->send();
+    }
 
     function delRowAction() {
 
         $dataManager = new Ui\DataManager($_GET['modelName'], $_GET['managerName']);
         if ($dataManager->chackAccess()) {
             $model = $_GET['modelName']::get($_GET['key'], $_GET['modelName']::index(), !empty($_GET['params']) ? $_GET['params'] : []);
+            if ($model) {
+                $model->delete(!empty($_GET['params']) ? $_GET['params'] : []);
+            }
+        }
+        $result = new Server\Result();
+        $result->send();
+    }
+    function delCategoryAction() {
+
+        $dataManager = new Ui\DataManager($_GET['modelName'], $_GET['managerName']);
+        if ($dataManager->chackAccess() && !empty($dataManager->managerOptions['categorys'])) {
+            $categoryModel = $dataManager->managerOptions['categorys']['model'];
+            $model = $categoryModel::get($_GET['key'], $categoryModel::index(), !empty($_GET['params']) ? $_GET['params'] : []);
             if ($model) {
                 $model->delete(!empty($_GET['params']) ? $_GET['params'] : []);
             }
